Add user-checking to document views

+++ api/api/views/documents.py
@@ -10,7 +10,6 @@ blueprint = Blueprint('document', __name__, url_prefix='/documents')
 document_schema = schemas.DocumentSchema()
 documents_schema = schemas.DocumentSchema(many=True)
 
-
 # Document CREATE
 @blueprint.route("", methods=['POST'])
 @jwt.require_jwt
@@ -43,10 +42,18 @@ def get_document(doc_id):
 @blueprint.route("/<int:doc_id>", methods=['PUT'])
 @jwt.require_jwt
 def edit_document(doc_id):
+    user = api.helpers.get_user()
     data = document_schema.load(request.get_json())
     doc = Document.query.get_or_404(doc_id)
+    xhr = MakeResponse()
+
+    if not doc.user_has_access(user, 'write'):
+        xhr.set_error(401, "Not authorized to edit document.")
+        return xhr.response
+
     for k, v in data.data.iteritems():
         setattr(doc, k, v)
+
     doc.save()
     xhr = MakeResponse(200, body=document_schema.dump(doc).data)
     return xhr.response
@@ -55,8 +62,13 @@ def edit_document(doc_id):
 @blueprint.route("/<int:doc_id>", methods=['DELETE'])
 @jwt.require_jwt
 def delete_document(doc_id):
+    user = api.helpers.get_user()
     doc = Document.query.get_or_404(doc_id)
+    xhr = MakeResponse(200)
+
+    if not doc.user_has_access(user, 'edit'):
+        xhr.set_error(401, "Not authorized to delete document.")
+        return xhr.response
+
     doc.delete()
-    xhr = MakeResponse()
-    xhr.set_status(200)
     return xhr.response
